app/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic-settings==2.1.0
twilio==8.13.0
python-dotenv==1.0.1
httpx==0.26.0
aioredis==2.0.1
supabase==2.3.7
pydub==0.25.1
google-cloud-aiplatform==1.41.0
google-api-python-client==2.117.0
Jinja2==3.1.3
typing-extensions
starlette
typing_extensions